Writing, Comment 추가 서비스도 변경

+++ src/main/java/com/sparta/spring_basic/service/WritingService.java
@@ -1,5 +1,7 @@
 package com.sparta.spring_basic.service;
 
+import static com.sparta.spring_basic.ExceptionHandler.ErrorCode.*;
+import com.sparta.spring_basic.dto.ResponseDto;
 import com.sparta.spring_basic.dto.WritingRequestDto;
 import com.sparta.spring_basic.entity.Writing;
 import com.sparta.spring_basic.repository.WritingRepository;
@@ -8,8 +10,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
-import java.net.PasswordAuthentication;
-import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -17,43 +18,66 @@ public class WritingService {
     private final WritingRepository writingRepository;
 
     //게시글 작성
-    public Writing writePost(WritingRequestDto requestDto){
-        Writing writing = new Writing(requestDto);  //받은 Dto로 entity 생성
-        return writingRepository.save(writing);
+    @Transactional
+    public ResponseDto<?> writePost(WritingRequestDto requestDto, String username){
+        Writing writing = new Writing(requestDto, username);  //받은 Dto로 entity 생성
+        writingRepository.save(writing);
+        return ResponseDto.success(writing);
     }
 
     //전체 게시글 조회
-    public List<Writing> getAllPosts() {
-        return writingRepository.findAll(Sort.by(Sort.Direction.DESC, "createdAt"));
+    public ResponseDto<?> getAllPosts() {
+        return ResponseDto.success(writingRepository.findAll(Sort.by(Sort.Direction.DESC, "createdAt")));
     }
 
     //id로 게시글 조회
-    public Writing getPost(Long id){
-        return writingRepository.findById(id).orElseThrow(
-                () -> new IllegalArgumentException("아이디가 존재하지 않습니다.")
-        );
+    public ResponseDto<?> getPost(Long id){
+        Optional<Writing> optionalWriting = writingRepository.findById(id);
+        if(optionalWriting.isEmpty()){
+            return ResponseDto.fail(NO_POST);
+        }
+
+        return ResponseDto.success(optionalWriting.get());
     }
 
-    public boolean checkPW(Long id, String password) {
-        Writing writing = writingRepository.findById(id).orElseThrow(
-                () -> new IllegalArgumentException("아이디가 존재하지 않습니다.")
-        );
-        return writing.getPassword().equals(password);
-    }
+//    public boolean checkPW(Long id, String password) {
+//        Writing writing = writingRepository.findById(id).orElseThrow(
+//                () -> new IllegalArgumentException("아이디가 존재하지 않습니다.")
+//        );
+//        return writing.getPassword().equals(password);
+//    }
 
     //게시글 수정
     @Transactional
-    public Writing updatePost(Long id, WritingRequestDto requestDto){
-        Writing writing = writingRepository.findById(id).orElseThrow(
-                () -> new IllegalArgumentException("아이디가 존재하지 않습니다.")
-        );
+    public ResponseDto<?> updatePost(Long id, WritingRequestDto requestDto, String username){
+        Optional<Writing> optionalWriting = writingRepository.findById(id);
+        if(optionalWriting.isEmpty()){
+            return ResponseDto.fail(NO_POST);
+        }
+
+        Writing writing = optionalWriting.get();
+        if(!writing.getAuthor().equals(username)){
+            return ResponseDto.fail(NO_AUTH);
+        }
+
         writing.update(requestDto);
-        return writing;
+        return ResponseDto.success(writing);
     }
 
     //게시글 삭제
-    public Long deletePost(Long id){
+    @Transactional
+    public ResponseDto<?> deletePost(Long id, String username){
+        Optional<Writing> optionalWriting = writingRepository.findById(id);
+        if(optionalWriting.isEmpty()){
+            return ResponseDto.fail(NO_POST);
+        }
+
+        Writing writing = optionalWriting.get();
+        if(!writing.getAuthor().equals(username)){
+            return ResponseDto.fail(NO_AUTH);
+        }
+
         writingRepository.deleteById(id);
-        return id;
+        return ResponseDto.success(id);
     }
 }
